+++ test/moduletest/vcodec/hwdecoder/include/videodec_ndk_sample.h
@@ -60,7 +60,7 @@ public:
     bool SURFACE_OUTPUT = false;
     uint32_t DEFAULT_WIDTH = 1920;
     uint32_t DEFAULT_HEIGHT = 1080;
-    uint32_t DEFAULT_FRAME_RATE = 30;
+    double DEFAULT_FRAME_RATE = 30.0;
     bool BEFORE_EOS_INPUT = false;              // 0800 测试用例
     bool BEFORE_EOS_INPUT_INPUT = false;        // 0900 测试用例
     bool AFTER_EOS_DESTORY_CODEC = true;        // 1000 测试用例 结束不销毁codec
